refactor(contracts): separate edit lowering helpers
Why:
- Ultra review found that edit-step lowering semantics were still mixed into the ContractReading namespace.
- The operation contract read policy also exposed an unused object-level flag while the reader only consumes field strictness.
- The moved helper tests were still collected in one broad class, making reader-specific regressions harder to localize.

What Changed:
- Moved edit-step lowering helpers into MackySoft.Ucli.Contracts.Ipc.EditSteps and documented the boundary.
- Removed the unused RequireOperationObject flag from OperationContractReadPolicy.
- Kept scene query resolution on a Unity-local QueryTargetKind and convert to IpcEditTargetKind only at edit compilation.
- Added architecture coverage to keep Unity operation phases from depending on contract edit-step lowering helpers.
- Split contract reading helper tests by reader and policy responsibility.

Impact:
- Public contracts and IPC JSON remain unchanged.
- Internal placement now separates JSON reading, edit-step lowering, and Unity scene query result categories.
