+++ src/Ucli.Contracts/Ipc/EditSteps/IpcEditTargetKind.cs
@@ -1,4 +1,4 @@
-namespace MackySoft.Ucli.Contracts.Ipc.ContractReading;
+namespace MackySoft.Ucli.Contracts.Ipc.EditSteps;
 
 /// <summary> Defines primitive target categories used during edit-step lowering. </summary>
 internal enum IpcEditTargetKind
