+++ tests/Ucli.Architecture.Tests/Architecture/PublicSurfaceBoundaryTests.cs
@@ -11,6 +11,7 @@ public sealed class PublicSurfaceBoundaryTests
             "MackySoft.Ucli.Application.Features",
             "MackySoft.Ucli.Application.Shared",
             "MackySoft.Ucli.Contracts.Ipc.ContractReading",
+            "MackySoft.Ucli.Contracts.Ipc.EditSteps",
             "MackySoft.Ucli.Features",
             "MackySoft.Ucli.Hosting",
             "MackySoft.Ucli.Shared",
